cookbook/core/containerization/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
absl-py==1.0.0
    # via
    #   tensorboard
    #   tensorflow-macos
arrow==1.2.2
    # via jinja2-time
astunparse==1.6.3
    # via tensorflow-macos
binaryornot==0.4.4
    # via cookiecutter
cachetools==5.0.0
    # via google-auth
certifi==2021.10.8
    # via requests
chardet==4.0.0
    # via binaryornot
charset-normalizer==2.0.12
    # via requests
checksumdir==1.2.0
    # via flytekit
click==8.1.2
    # via
    #   cookiecutter
    #   flytekit
cloudpickle==2.0.0
    # via flytekit
cookiecutter==1.7.3
    # via flytekit
croniter==1.3.4
    # via flytekit
cycler==0.11.0
    # via matplotlib
dataclasses-json==0.5.7
    # via flytekit
decorator==5.1.1
    # via retry
deprecated==1.2.13
    # via flytekit
diskcache==5.4.0
    # via flytekit
docker-image-py==0.1.12
    # via flytekit
docstring-parser==0.13
    # via flytekit
flatbuffers==2.0
    # via tensorflow-macos
flyteidl==0.24.20
    # via flytekit
flytekit==0.32.0
    # via -r ../../common/requirements-common.in
fonttools==4.32.0
    # via matplotlib
gast==0.5.3
    # via tensorflow-macos
google-auth==2.6.3
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow-macos
googleapis-common-protos==1.56.0
    # via
    #   flyteidl
    #   grpcio-status
grpcio==1.44.0
    # via
    #   flytekit
    #   grpcio-status
    #   tensorboard
    #   tensorflow-macos
grpcio-status==1.44.0
    # via flytekit
h5py==3.6.0
    # via tensorflow-macos
idna==3.3
    # via requests
importlib-metadata==4.11.3
    # via
    #   keyring
    #   markdown
jinja2==3.1.1
    # via
    #   cookiecutter
    #   jinja2-time
jinja2-time==0.2.0
    # via cookiecutter
joblib==1.1.0
    # via scikit-learn
keras==2.8.0
    # via tensorflow-macos
keras-preprocessing==1.1.2
    # via tensorflow-macos
keyring==23.5.0
    # via flytekit
kiwisolver==1.4.2
    # via matplotlib
libclang==13.0.0
    # via tensorflow-macos
markdown==3.3.6
    # via tensorboard
markupsafe==2.1.1
    # via jinja2
marshmallow==3.15.0
    # via
    #   dataclasses-json
    #   marshmallow-enum
    #   marshmallow-jsonschema
marshmallow-enum==1.5.1
    # via dataclasses-json
marshmallow-jsonschema==0.13.0
    # via flytekit
matplotlib==3.5.1
    # via -r ../../common/requirements-common.in
mypy-extensions==0.4.3
    # via typing-inspect
natsort==8.1.0
    # via flytekit
numpy==1.22.3
    # via
    #   h5py
    #   keras-preprocessing
    #   matplotlib
    #   opt-einsum
    #   pandas
    #   pyarrow
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   tensorflow-macos
oauthlib==3.2.0
    # via requests-oauthlib
opt-einsum==3.3.0
    # via tensorflow-macos
packaging==21.3
    # via
    #   marshmallow
    #   matplotlib
pandas==1.4.2
    # via
    #   -r requirements.in
    #   flytekit
pillow==9.1.0
    # via matplotlib
poyo==0.5.0
    # via cookiecutter
protobuf==3.20.0
    # via
    #   flyteidl
    #   flytekit
    #   googleapis-common-protos
    #   grpcio-status
    #   protoc-gen-swagger
    #   tensorboard
    #   tensorflow-macos
protoc-gen-swagger==0.1.0
    # via flyteidl
py==1.11.0
    # via retry
pyarrow==6.0.1
    # via flytekit
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyparsing==3.0.8
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   arrow
    #   croniter
    #   flytekit
    #   matplotlib
    #   pandas
python-json-logger==2.0.2
    # via flytekit
python-slugify==6.1.1
    # via cookiecutter
pytimeparse==1.1.8
    # via flytekit
pytz==2022.1
    # via
    #   flytekit
    #   pandas
pyyaml==6.0
    # via flytekit
regex==2022.3.15
    # via docker-image-py
requests==2.27.1
    # via
    #   cookiecutter
    #   flytekit
    #   requests-oauthlib
    #   responses
    #   tensorboard
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
responses==0.20.0
    # via flytekit
retry==0.9.2
    # via flytekit
rsa==4.8
    # via google-auth
scikit-learn==1.0.2
    # via sklearn
scipy==1.8.0
    # via scikit-learn
six==1.16.0
    # via
    #   absl-py
    #   astunparse
    #   cookiecutter
    #   google-auth
    #   google-pasta
    #   grpcio
    #   keras-preprocessing
    #   python-dateutil
    #   tensorflow-macos
sklearn==0.0
    # via -r requirements.in
sortedcontainers==2.4.0
    # via flytekit
statsd==3.3.0
    # via flytekit
tensorboard==2.8.0
    # via tensorflow-macos
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
tensorflow-macos==2.8.0
    # via -r requirements.in
termcolor==1.1.0
    # via tensorflow-macos
text-unidecode==1.3
    # via python-slugify
tf-estimator-nightly==2.8.0.dev2021122109
    # via tensorflow-macos
threadpoolctl==3.1.0
    # via scikit-learn
typing-extensions==4.1.1
    # via
    #   flytekit
    #   tensorflow-macos
    #   typing-inspect
typing-inspect==0.7.1
    # via dataclasses-json
urllib3==1.26.9
    # via
    #   flytekit
    #   requests
    #   responses
werkzeug==2.1.1
    # via tensorboard
wheel==0.37.1
    # via
    #   -r ../../common/requirements-common.in
    #   astunparse
    #   flytekit
    #   tensorboard
wrapt==1.14.0
    # via
    #   deprecated
    #   flytekit
    #   tensorflow-macos
zipp==3.8.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
